Assignment1 solutions
