--- src/AssignmentSolutions/InsufficientFundsException.java
@@ -1,7 +1,0 @@
-package AssignmentSolutions;
-
-public class InsufficientFundsException extends Exception {
-    public InsufficientFundsException(String message) {
-        super(message);
-    }
-}
