+++ src/Assignment_Solutions/InsufficientFundsException.java
@@ -1,0 +1,7 @@
+package Assignment_Solutions;
+
+public class InsufficientFundsException extends Exception {
+    public InsufficientFundsException(String message) {
+        super(message);
+    }
+}
